+++ src/main/java/nl/qien/uren/repository/UserRepository.java
@@ -13,5 +13,5 @@ public interface UserRepository {
 
     List<User> findAll();
 
-    Long getMaxId();
+    int getMaxId();
 }
